fix: remove infra from server cascade and split oversized test file
Remove 'infra' from LAYER_CASCADE_CONSERVATIVE 'server' entry —
server changes should not trigger full infra dir; the 9 unconditional
infra files are included via direct_test_files by the tiered logic.

Move AC3/AC4 file-level scope tests to test_test_filter_scope_extras.py
to bring test_test_filter.py under the 1000-line limit.
